settingsview

+++ ChickenInvasionlibGDX/core/src/com/chicken/invasion/GameButton.java
@@ -33,7 +33,7 @@ public class GameButton{
         this.sprite.setY(y);
     }
 
-    public void setSize(int w, int h){
+    public void setSize(float w, float h){
         this.sprite.setSize(w, h);
     }
